apimeuspedidos/requirements.txt
django==1.8.5
djangorestframework
markdown
django-filter
